[ET-590] Initial SDK with support for connecting to CCT system to create a campaign
This initial support includes the ability to start a campaign, add media for the campaign, set campaign data, and place the campaign. It also includes the ability to list the customers that the access token has access to, organized by agency. Additionally, it allows for the listing of campaign flows that a client has access to.