Add endpoints for pause campaign, resume campaign and get campaign state
We have implemented three new endpoints to provide campaign state information and to enable the pausing or resuming of campaigns. Please note that the pausing or resuming of campaigns is currently only available for those campaigns sent to the Monolith platform, and attempting to pause or resume a campaign on the Campaign Wizard system will result in a 400 response. However, we are planning to add support for this functionality on the Campaign Wizard system in the near future.